Update - consolidation
Consolidation of information, small restructure changes (moved contribute to main page with intention of attracting more developers to the wiki), update to file names for developers for readability, added educational resources for Learn Tezos category with wallet information, update to links in DEX resources, NFT resources, Game resources. Alphabetical organization of links for  block explorer, and others. Update to file names for developer ease of readability. Organization of files within the categories to be more fluid and easier to use and read.